% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itan.R
\name{itan}
\alias{itan}
\title{item analysis}
\usage{
itan(x, keys = NULL, digits = 3, no.resp = 4)
}
\arguments{
\item{x}{x is an n subjects by k items matrix}

\item{keys}{is a vector with the correct keys (A, B, C, D)}
}
\description{
item analysis
}
\author{
Christophe Lalanne <https://gist.github.com/chlalanne>
}

