% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ittReg.R
\name{ittReg}
\alias{ittReg}
\title{Intent to treat regression}
\usage{
ittReg(y = outcome, tx = tx.assign, x = covariates, data = df, ...)
}
\arguments{
\item{y}{outcome, default is outcome}

\item{tx}{treatment assignment variable}

\item{x}{covariates, default named covariates}

\item{data}{data frame, default named df}
}
\description{
Intent to treat regression
}
\author{
William Murrah
}

