% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplus.get.group.dataset.R
\name{mplus.get.group.dataset}
\alias{mplus.get.group.dataset}
\title{supporting function for getting dataset}
\usage{
mplus.get.group.dataset(file, groupstr, datastr)
}
\arguments{
\item{file}{the quoted name of an existing GH5 file}

\item{groupstr}{the name of the group for the attribute}

\item{data}{sr the name of the attribute}
}
\description{
mplus.get.group.dataset('ex8.1.gh5','bayesian_data','statements')
}
\author{
Thuy Nguyen, Muthen & Muthen
}

