% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplus_view_plots.R
\name{mplus_view_plots}
\alias{mplus_view_plots}
\title{View Mplus plots in GH5 file}
\usage{
mplus_view_plots(file)
}
\arguments{
\item{file}{}
}
\description{
View Mplus plots in GH5 file
}
\author{
Thuy Nguyen, Muthen & Muthen
}

