% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newR.R
\name{newR}
\alias{newR}
\title{Create new R script}
\usage{
newR(filename, author = "William Murrah", subdir = "R/", open = TRUE)
}
\arguments{
\item{filename}{Name of new script file}

\item{author}{Name of author, default is "William Murrah"}

\item{subdir}{Subdirectory to place file, default, R/}

\item{open}{logical, whether to open script, default TRUE}
}
\description{
Create new R script
}
