% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryR.R
\name{summaryR}
\alias{summaryR}
\title{summaryR}
\usage{
summaryR(model, type = c("hc3", "hc0", "hc1", "hc2", "hc4"), ...)
}
\arguments{
\item{model}{from lm()}

\item{type}{see hccm() in car package}
}
\description{
summaryR
}
\author{
John Fox, tony91782 <https://gist.github.com/tony91782/954316>
}
