% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/totReg.R
\name{totReg}
\alias{totReg}
\title{Creates an treatment on the treated model with the ivreg function from
the AER package.}
\usage{
totReg(y = outcome, iv = instrument, tr = treatment, x = covariates,
  data = df, ...)
}
\arguments{
\item{y}{Character vector of length = 1 that is the column name of the
outcome variable.}

\item{iv}{Character vector of length = 1 that is the column name of the
insrumental variable (e.g. treatment assignment).}

\item{tr}{Character vector of length = 1 that is the column name of the
treatment received variable.}

\item{x}{Character vector of any length that contains the column names of
the covariates.}

\item{data}{The data frame containing the variables referenced by the previous
arguments.}
}
\value{
An object of class 'ivreg' containing elements of model objects similar
  to the lm() function.
}
\description{
Creates an treatment on the treated model with the ivreg function from
the AER package.
}
\author{
William Murrah
}

