% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplus_get_file_dataset.R
\name{mplus_get_file_dataset}
\alias{mplus_get_file_dataset}
\title{mplus_get_file_dataset}
\usage{
mplus_get_file_dataset(file, datastr)
}
\arguments{
\item{file}{}

\item{datastr}{}
}
\description{
mplus_get_file_dataset
}
\author{
Thuy Nguyen, Muthen & Muthen
}

